Add modules, Update-list with changes, and clear-list functionalities
