+++ src/modules/todo.js
@@ -1,15 +1,6 @@
-const getFromLocalStorage = () => {
-  let data = JSON.parse(localStorage.getItem('todos'));
-  if (!data) {
-    data = [];
-  }
-  return data;
-};
+import { getFromLocalStorage, saveToLocalStorage } from './local-storage.js';
 
-const saveToLocalStorage = (todos) => {
-  localStorage.setItem('todos', JSON.stringify(todos));
-};
-
+// Adds new TODO item to the list
 const addTodo = () => {
   const todos = getFromLocalStorage();
   const newTodoDescription = document.getElementById('todo-input');
@@ -23,12 +14,14 @@ const addTodo = () => {
   saveToLocalStorage(todos);
 };
 
+// Updates single TODO item after edit
 const updateTodo = (todos, index, value) => {
   todos[index].description = value;
 
   saveToLocalStorage(todos);
 };
 
+// Deletes single TODO from the list
 const deleteTodoItem = (todos, index) => {
   const newTodos = todos.filter((todo, innerIndex) => index !== innerIndex);
   for (let i = 0; i < newTodos.length; i += 1) {
@@ -37,6 +30,7 @@ const deleteTodoItem = (todos, index) => {
   saveToLocalStorage(newTodos);
 };
 
+// Deletes all completed Items in the list
 const deleteAllCompleted = (todos) => {
   const newTodos = todos.filter((todo) => todo.completed !== true);
   for (let i = 0; i < newTodos.length; i += 1) {
